Move core/bids.py to core/construct_bids.py
The previous name was interfering with the import of bids from pybids in
select situations
