+++ snakebids/core/__init__.py
@@ -2,7 +2,7 @@
 __submodules__ = ["bids", "filtering", "input_generation"]
 
 # <AUTOGEN_INIT>
-from snakebids.core.bids import bids, print_boilerplate
+from snakebids.core.construct_bids import bids, print_boilerplate
 from snakebids.core.filtering import filter_list, get_filtered_ziplist_index
 from snakebids.core.input_generation import (
     generate_inputs,
